+++ src/lib.rs
@@ -401,218 +401,326 @@ mod tests {
         // -0.5
         let f = I0F32::from_bits(-1 << 31);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I0F32::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I0F32::from_num(0), true));
         assert_eq!(f.overflowing_round(), (I0F32::from_num(0), true));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I0F32::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I0F32::from_num(0), false)
+        );
 
         // -0.5 + Δ
         let f = I0F32::from_bits((-1 << 31) + 1);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I0F32::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I0F32::from_num(0), true));
         assert_eq!(f.overflowing_round(), (I0F32::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I0F32::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I0F32::from_num(0), false)
+        );
 
         // 0.5 - Δ
         let f = I0F32::from_bits((1 << 30) - 1 + (1 << 30));
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I0F32::from_num(0), true));
         assert_eq!(f.overflowing_floor(), (I0F32::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I0F32::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I0F32::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I0F32::from_num(0), false)
+        );
 
         // -0.5 - Δ
         let f = I1F31::from_bits(((-1) << 30) - 1);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(-1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(-1), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(-1), false)
+        );
 
         // -0.5
         let f = I1F31::from_bits((-1) << 30);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(-1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(0), false)
+        );
 
         // -0.5 + Δ
         let f = I1F31::from_bits(((-1) << 30) + 1);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(0), false)
+        );
 
         // 0.5 - Δ
         let f = I1F31::from_bits((1 << 30) - 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(-1), true));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(0), false)
+        );
 
         // 0.5
         let f = I1F31::from_bits(1 << 30);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(-1), true));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(-1), true));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(0), false)
+        );
 
         // 0.5 + Δ
         let f = I1F31::from_bits((1 << 30) + 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I1F31::from_num(-1), true));
         assert_eq!(f.overflowing_floor(), (I1F31::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I1F31::from_num(-1), true));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I1F31::from_num(-1), true));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I1F31::from_num(-1), true)
+        );
 
         // -3.5 - Δ
         let f = I16F16::from_bits(((-7) << 15) - 1);
         assert_eq!(f.to_num::<i32>(), -4);
+        assert_eq!(f.round_to_zero(), -3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-4), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-4), false)
+        );
 
         // -3.5
         let f = I16F16::from_bits((-7) << 15);
         assert_eq!(f.to_num::<i32>(), -4);
+        assert_eq!(f.round_to_zero(), -3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-4), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-4), false)
+        );
 
         // -3.5 + Δ
         let f = I16F16::from_bits(((-7) << 15) + 1);
         assert_eq!(f.to_num::<i32>(), -4);
+        assert_eq!(f.round_to_zero(), -3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-4), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-3), false)
+        );
 
         // -2.5 - Δ
         let f = I16F16::from_bits(((-5) << 15) - 1);
         assert_eq!(f.to_num::<i32>(), -3);
+        assert_eq!(f.round_to_zero(), -2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-2), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-3), false)
+        );
 
         // -2.5
         let f = I16F16::from_bits((-5) << 15);
         assert_eq!(f.to_num::<i32>(), -3);
+        assert_eq!(f.round_to_zero(), -2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-2), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-2), false)
+        );
 
         // -2.5 + Δ
         let f = I16F16::from_bits(((-5) << 15) + 1);
         assert_eq!(f.to_num::<i32>(), -3);
+        assert_eq!(f.round_to_zero(), -2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(-2), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-2), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-2), false)
+        );
 
         // -0.5 - Δ
         let f = I16F16::from_bits(((-1) << 15) - 1);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(-1), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(-1), false)
+        );
 
         // -0.5
         let f = I16F16::from_bits((-1) << 15);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(-1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(0), false)
+        );
 
         // -0.5 + Δ
         let f = I16F16::from_bits(((-1) << 15) + 1);
         assert_eq!(f.to_num::<i32>(), -1);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(-1), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(0), false)
+        );
 
         // 0.5 - Δ
         let f = I16F16::from_bits((1 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(0), false)
+        );
 
         // 0.5
         let f = I16F16::from_bits(1 << 15);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(0), false)
+        );
 
         // 0.5 + Δ
         let f = I16F16::from_bits((1 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(1), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(1), false)
+        );
 
         // 2.5 - Δ
         let f = I16F16::from_bits((5 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(2), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(2), false)
+        );
 
         // 2.5
         let f = I16F16::from_bits(5 << 15);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(2), false)
+        );
 
         // 2.5 + Δ
         let f = I16F16::from_bits((5 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(3), false)
+        );
 
         // 3.5 - Δ
         let f = I16F16::from_bits((7 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(3), false)
+        );
 
         // 3.5
         let f = I16F16::from_bits(7 << 15);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(4), false)
+        );
 
         // 3.5 + Δ
         let f = I16F16::from_bits((7 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (I16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (I16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (I16F16::from_num(4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (I16F16::from_num(4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (I16F16::from_num(4), false)
+        );
     }
 
     #[test]
@@ -620,97 +728,145 @@ mod tests {
         // 0.5 - Δ
         let f = U0F32::from_bits((1 << 31) - 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U0F32::from_num(0), true));
         assert_eq!(f.overflowing_floor(), (U0F32::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U0F32::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U0F32::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U0F32::from_num(0), false)
+        );
 
         // 0.5
         let f = U0F32::from_bits(1 << 31);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U0F32::from_num(0), true));
         assert_eq!(f.overflowing_floor(), (U0F32::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U0F32::from_num(0), true));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U0F32::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U0F32::from_num(0), false)
+        );
 
         // 0.5 + Δ
         let f = U0F32::from_bits((1 << 31) + 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U0F32::from_num(0), true));
         assert_eq!(f.overflowing_floor(), (U0F32::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U0F32::from_num(0), true));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U0F32::from_num(0), true));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U0F32::from_num(0), true)
+        );
 
         // 0.5 - Δ
         let f = U16F16::from_bits((1 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(0), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(0), false)
+        );
 
         // 0.5
         let f = U16F16::from_bits(1 << 15);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(0), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(0), false)
+        );
 
         // 0.5 + Δ
         let f = U16F16::from_bits((1 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 0);
+        assert_eq!(f.round_to_zero(), 0);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(1), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(0), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(1), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(1), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(1), false)
+        );
 
         // 2.5 - Δ
         let f = U16F16::from_bits((5 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(2), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(2), false)
+        );
 
         // 2.5
         let f = U16F16::from_bits(5 << 15);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(2), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(2), false)
+        );
 
         // 2.5 + Δ
         let f = U16F16::from_bits((5 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 2);
+        assert_eq!(f.round_to_zero(), 2);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(2), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(3), false)
+        );
 
         // 3.5 - Δ
         let f = U16F16::from_bits((7 << 15) - 1);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(3), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(3), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(3), false)
+        );
 
         // 3.5
         let f = U16F16::from_bits(7 << 15);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(4), false)
+        );
 
         // 3.5 + Δ
         let f = U16F16::from_bits((7 << 15) + 1);
         assert_eq!(f.to_num::<i32>(), 3);
+        assert_eq!(f.round_to_zero(), 3);
         assert_eq!(f.overflowing_ceil(), (U16F16::from_num(4), false));
         assert_eq!(f.overflowing_floor(), (U16F16::from_num(3), false));
         assert_eq!(f.overflowing_round(), (U16F16::from_num(4), false));
-        assert_eq!(f.overflowing_round_ties_to_even(), (U16F16::from_num(4), false));
+        assert_eq!(
+            f.overflowing_round_ties_to_even(),
+            (U16F16::from_num(4), false)
+        );
     }
 }